Set up adding a cable to the table
Adds a single row to the table, just to hold the place of a cable item. Proceeds to pane for setting up type and quantity.
Alerts user if they attempt to add a cable without finishing the current one. Cable index is stored on cookie. Currently untested against adding after a complete cable. Should still work...